sklearn-pipeline/requirements.txt
pandas
scikit-learn
joblib
hydra-core
omegaconf